Added a DietForm component to separate the form logic and UI and updated AddDietScreen to use DietForm component for adding new diet entry.
